corrected type

+++ src/main/java/it/unibo/cloudnine/data/Menu.java
@@ -4,6 +4,6 @@ package it.unibo.cloudnine.data;
  * A single instance of a menu with a name, a cost and a certain 
  * number of plates.
  */
-public record Menu(String nome, float costo, int numPiatti) {
+public record Menu(String nome, float costo, long numPiatti) {
 
 }
